Re-add error reporting (throwing) support
Error reporting support was dropped with the migration to Duktape, since
there is no direct equivalent of mozjs's JS_ReportError() function.
However, it's easy to implement something similar using duk_throw().

Let's add our own implementation based on duk_throw() and re-add the
macros around this function.

WARNING: The behavior is similar to mozjs but not identical. Because
duk_throw() never returns, our wrappers do not return either. By
contrast, JS_ReportError() returns, and the native C function that calls
it must return -1 for the error to be thrown. This means that code like
the example below will NOT work as expected:

	ret = js_report_error(ctx, "This is an error");
	// do some cleanup here
	return ret;

More specifically, the cleanup code WILL NOT RUN in the example above.